Move Astoria logic to own modules
